refactor: :art: import firebase type from config file
