+++ src/index.ts
@@ -2,8 +2,7 @@ import cors from 'cors'
 import 'dotenv/config'
 import express from 'express'
 import 'express-async-errors'
-import { PORT } from './config/const'
-import { corsOption } from './config/cors'
+import { PORT, corsOption } from './config'
 import authMiddleware from './middlewares/Auth'
 import errorHandlerMiddleware from './middlewares/ErrorHandler'
 import redirectMiddleware from './middlewares/Redirect'
